v0.2.4 Maintenance Release
Sometimes BrowserStack tunnels get stopped and since the behavior
is inconsistent, there is no straightforward solution. As of now,
a retry mechanism has been added to Native Runner, where a setting
'retries' can be added to cbtr.json (default being 0), and for each
test that fails, it is tried that many times.

[ci skip]