v0.5.3 Maintenance Release
Fixes #27

SauceLabs provides us with short version and long version for browsers;
however, we add trailing 0s to work well with our browser conf files.
This goes OK with most browsers except SauceLabs where the trailing 0
creates a version not found error. Ideally, it has to be the pure short
version (which won't go well with our conf files), or the long version
(which makes users' yml files clumsy and we can't use our smart compact
format anymore in several cases). Figured out that since the problem is
only for Microsoft Edge, a combination of using the short version and
long version solves things for us.

Fixes #22

With native runner rewrite for parallel builds, the logic was lost.
Now since we are looking to use runMultiple and runScriptMultiple,
the place to figure out session name with uuidv4 goes out of all
bin/server layers all the way to core Platform layer, because it
has to be done for each job/scriptjob separately.

Eventually, testem's logic that does it inside its hooks was removed.
Now core.utils has a function that provides the build parameters
derivation. And in core.Job and core.ScriptJob this function gets
called in create/createMultiple.

Verified with SauceLabs and BrowserStack JS and Selenium runs as
well as Testem functional tests.

Fixes #26

The closest point to intercept the thrown error is the point where
Bluebird.all promise of starting all tunnels is returned. It now has
a following catch block to ensure the rejection does not become
unhandled.

Fixes #25

There is a race condition between test.stop() and native runner
monitor.

Our stop implementation ensures that after a DELETE call, the status
returned for a given job does say 'stopped', which happens with retries.
The native runner monior is sneaking through this period to see twice
that status is stopped but native runner has not marked it as stopped
yet.

The solution used here is that as soon as the test results arrive and
we find a valid non-stopped test to stop, a state 'nativeRunnerStopping'
is added to the test. When native runner monitor sees this state, it
just won't process this test any more to figure out if it has been not
reported yet etc. So for actually failing tests that do not respond back,
the earlier logic (now encapsulated under the if condition checking for
this state) comes into play, and would determine correctly, free of the
race conditions, that the test did not stop correctly.

Bumped up version to 0.5.3

Also modified the acknowledgement logos

[ci skip]