v0.5.4 Maintenance Release
Fixes #12

Code coveage of each executable binary does not seem possible on
Windows. Wrapping the executable to cover under istanbul cover did
not work because on Windows the JS files do not run as standalone
executables, a known issue/behavior. So the only way it could work
is where we could run 'node istanbul cover node -- executable args';
but this fails on Linux/OSX itself. This is a dead end as of now.

As of now, the bin/ coverage hence would be limited on Windows.

conf/ updates:
- High Sierra now on BrowserStack. Conf updated.

build issues:
- mocha 4.x seems to display a hang as seen with issues filed very
  recently. Out of the 3 packages changed since last build, mocha
  was the only one with a major version jump. Downgraded to 3.5.3

localhost to 127.0.0.1

On Appveyor, the single scriptjob test accessing 'localhost' failed
with bad gateway error and the browser could not resolve the domain
name 'localhost'.

For safety sake, the direct IP address 127.0.0.1 is now used in all
tests.

Build Failures Fixes

Supported browsers/platforms have changed breaking one of the functional
tests for cbtr-init. Fixed.

The above was the only reason of failure on Travis. However; on Appveyor
there were several tunnel failures. It looked like that it could be something
to do with change to 8.6.0 (automatically through the stable alias) from 8.5.0;
but the most tunnel failures were with 4.8.0. So it may be that it has
something to do with the Appveyor environment.

README updated

[ci skip]

Create CODE_OF_CONDUCT.md

[ci skip]

Create CONTRIBUTING.md

[ci skip]

Create ISSUE_TEMPLATE.md

[ci skip]

Bumped up package version

[ci skip]

Changed keywords for the NPM package