v1.1.1 Minor Version Release
Travis Build Errors Fixes

- Having default port 8000 messes up native runner tests.
- Added libstdc++ package installation for Linux as required for
  CrossBrowserTesting tunnels
- BrowserStack and SauceLabs platform tests had syntax errors that
  got missed somehow
- SauceLabs scriptjob test has one of those failing where the page
  source is no longer about using Sauce Connect because of moving
  from 127.0.0.1 to build.cross-browser-tests-runner.org
- CrossBrowserTesting platform tests for messy status have some kind
  of race conditions, especially when it comes to ScriptJob runs.

- The earlier method used to get libstdc++ on Travis linux VMs failed
  with another error of linking. Using a simpler and clearer method
  to install the required libstdc++ package.

CrossBrowserTesting Linux blocker fix:

On Linux, the tunnel obtained from app.crossbrowsertesting.com link
(a .zip archive) does not work with various issues.

/usr/lib/x86_64-linux-gnu/libstdc++.so.6: version `GLIBCXX_3.4.20' not found

This happens with the default libstdc++6 version on Ubuntu Trusty.

relocation error: ./cbt-tunnels-linux: symbol _ZTVNSt7__cxx1115basic_stringbufIcSt11char_traitsIcESaIcEEE, version GLIBCXX_3.4.21 not defined in file libstdc++.so.6 with link time reference

This happens once PPA repositories are used to get latest libstdc++
possible on Trusty.

There are no further solutions except upgrading to Ubuntu 16.x or
compiling the tunnel from source code, both of them not possible in
our case.

An alternative exists, and that's the better alternative as per the
support team. This is to get the tunnels from their github at:

https://github.com/crossbrowsertesting/cbt-tunnel-nodejs/releases

The binaries available here behave a bit differently but mostly the
same functionality is achieved. The abstractions had to be changed
quite a bit because these are not ZIP files but just the individual
binaries. So eveywhere in lib and tests there was a change to Binary
from Archive.

Bug Fixes

- cbtr-init can now merge into/update existing settings file with
  the '--update|-u' flag. Fixes #16
- cbtr-init had a bug around version numbers. The extra stress on
  decimal numbers being version numbers is contrived, and is simplified.
- Server was broken because of bad path of conf files. It should be
  relative to cross-browser-tests-runner root and not project root.
  Fixed.